11:37 PM 10/21/2020
This file is the one I want to share if  I was working on a software. It has the same name of another . And it in a repo thats in a rep.

So I made a new text file using notepad from the Git CMD like so;

	start notepad examplefile.txt

and then i typed this into it. Why? cos I need to see if I can use the prompt to interact with GitHub from C:\Users\novas\GIT17ASAP

so there!