## Bioreaching
Environment Tobago hopes the GEF funded project Bioreach encounters no more stumbling blocks (such as a pandemic) before it hits Trinidad and Tobago ground. To prepare for its contribution your NGO is into the process of leasing a plot in the southwest Even though arrangements aren't concretised at time of writing (late October 2020) we have already begun work on to estate, getting to know our Prospects as it were. What are we doing? To report in a way that makes sense in a year when nothing does we must go back to April when the measures against the coronavirus took precedence over well everything.

Reacting in a positively if absolutely worried way we looked to getting food into the ground. Not so much as for whom but As soon as possible please and thank you. So said so done ET entered the Prospect. At this point it behooves us to say a heartfelt Gracias to Mr Gregory Weeks scion of the Prospect Estate who took to the idea of re-establishing his inheritance as a player in the Tobago food supply chain. And so began C20, a community garden that saw the first fork biting earth in the COVID year 2020.

As indicated up top we were, are aware the Bioreach mechanism is underway, under review also as the coronavirus forces changes to the way projects are implemented and monitored.